方案分享/赛题四/天池三轮车/docker_image/docker_image/app/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
pyecharts-snapshot
pyecharts~=2.0.8
requests == 2.31.0
pillow~=11.1.0
efinance>=0.5.0
akshare>=1.12.0
openai
selenium
duckduckgo-search>=3.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
ipython>=7.0.0
k-sogou-search
json-repair
mdformat
pypandoc
qwen_agent==0.0.27
loguru==0.7.3
qwen-agent[code_interpreter]